pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-price-prediction"
version = "0.1.0"
description = "Machine learning models for cryptocurrency price movement prediction"
authors = [{name = "Crypto ML Team"}]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "xgboost>=1.5.0",
    "tensorflow>=2.8.0",
    "lightgbm>=3.0.0",
    "PyYAML>=6.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "scipy>=1.7.0",
    "pandas-ta>=0.3.14b",
    "pytz>=2021.1",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "pytest>=6.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]
plotting = [
    "plotly>=5.0.0",
]
performance = [
    "numba>=0.56.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/crypto-price-prediction"
Repository = "https://github.com/yourusername/crypto-price-prediction"
Issues = "https://github.com/yourusername/crypto-price-prediction/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-dir]
"" = "."

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]